Fix warnings (#131)
* Fix -Wunused-parameter, start/end

* Fix -Wsign-compare in return

* Fix -Wsign-compare

The next definition:
using TreeNodeIndex = int;

, is compared with size_t

As it is only used like iterator, it will be better to put directly
size_t

* Fix -Wunknown-pragmas

Remove the pragma clang directives, with if defined(x) compiler
predirectives, when the compiler is not clang.

* Fix -Wsign-compare

Changed the input parameters by size_t, in order to compare same types

* Fix -Wmaybe-uninitialized

Set default values in the internal variables: 'fistParticle' and
'lastParticle'.

* Fix -Wsign-compare

Set numParticles like 'size_t', instead 'int', because is used in loops
where the index is compared with size_t.

* Fix -Wsign-compare

Comparation between 'unsigned int' and 'int' is done now as 'size_t' in
both sides.

* Fix include not found

Remove the include:
#include "../Octree.hpp"

, which is not in the code and not used anymore

* Fix code when is defined USE_OMP_TARGET

* Fix code when is defined USE_ACC

* Improve Makefile, including '-O3 -Wall -Wextra' in all the objectives.

It substituted the previous -O2 and add more Warning checks

* Improve Makefile including the PGI compiler headers location.

It is use when the objective 'mpi+omp+acc' is in use.

The OpenACC PGI support could be added with the package: 'NVIDIA
HPC-SDK' (https://developer.nvidia.com/hpc-sdk)

* Update domain/include/cstone/domain/domain.hpp


* Update changes in response to the comments in the PR.

Substitute TAB by spaces.

* Sebastian's suggestion in the PR.

Change size_t in the Task.size(), instead in the calls

+++ include/FileUtils.hpp
@@ -87,7 +87,7 @@ void writeParticleDataToBinFile(const std::string &path, Args &&... data)
 }
 
 template<typename... Args>
-void writeParticleDataToAsciiFile(int firstIndex, int lastIndex, const std::string& path, const bool append,
+void writeParticleDataToAsciiFile(size_t firstIndex, size_t lastIndex, const std::string& path, const bool append,
                                   const char separator, Args&&... data)
 {
     std::ios_base::openmode mode;
